move launchable types to pkg/launch from manifest
