+++ pkg/artifact/registry.go
@@ -7,7 +7,7 @@ import (
 	"net/url"
 
 	"github.com/square/p2/pkg/auth"
-	"github.com/square/p2/pkg/manifest"
+	"github.com/square/p2/pkg/launch"
 	"github.com/square/p2/pkg/osversion"
 	"github.com/square/p2/pkg/uri"
 	"github.com/square/p2/pkg/util"
@@ -25,7 +25,7 @@ type Registry interface {
 	// Given a LaunchableStanza from a pod manifest, returns a URL from which the
 	// artifact can be fetched an a struct containing the locations of files that
 	// can be used to verify artifact integrity
-	LocationDataForLaunchable(launchableID manifest.LaunchableID, stanza manifest.LaunchableStanza) (*url.URL, auth.VerificationData, error)
+	LocationDataForLaunchable(launchableID launch.LaunchableID, stanza launch.LaunchableStanza) (*url.URL, auth.VerificationData, error)
 }
 
 type registry struct {
@@ -59,7 +59,7 @@ func NewRegistry(registryURL *url.URL, fetcher uri.Fetcher, osVersionDetector os
 // manifest: ".manifest"
 // manifest signature: ".manifest.sig"
 // build signature: ".sig"
-func (a registry) LocationDataForLaunchable(launchableID manifest.LaunchableID, stanza manifest.LaunchableStanza) (*url.URL, auth.VerificationData, error) {
+func (a registry) LocationDataForLaunchable(launchableID launch.LaunchableID, stanza launch.LaunchableStanza) (*url.URL, auth.VerificationData, error) {
 	if stanza.Location == "" && stanza.Version.ID == "" {
 		return nil, auth.VerificationData{}, util.Errorf("Launchable must provide either \"location\" or \"version\" fields")
 	}
@@ -93,7 +93,7 @@ type RegistryResponse struct {
 	BuildSignatureLocation    string `json:"signature_location"`
 }
 
-func (a registry) fetchRegistryData(launchableID manifest.LaunchableID, version manifest.LaunchableVersion) (*url.URL, auth.VerificationData, error) {
+func (a registry) fetchRegistryData(launchableID launch.LaunchableID, version launch.LaunchableVersion) (*url.URL, auth.VerificationData, error) {
 	requestURL := &url.URL{
 		Path: fmt.Sprintf("%s/%s", discoverBasePath, launchableID),
 	}
