+++ pkg/hoist/hoist_launchable.go
@@ -14,7 +14,6 @@ import (
 	"github.com/square/p2/pkg/auth"
 	"github.com/square/p2/pkg/cgroups"
 	"github.com/square/p2/pkg/launch"
-	"github.com/square/p2/pkg/manifest"
 	"github.com/square/p2/pkg/p2exec"
 	"github.com/square/p2/pkg/runit"
 	"github.com/square/p2/pkg/user"
@@ -23,7 +22,7 @@ import (
 
 // A HoistLaunchable represents a particular install of a hoist artifact.
 type Launchable struct {
-	Id               manifest.LaunchableID // A (pod-wise) unique identifier for this launchable, used to distinguish it from other launchables in the pod
+	Id               launch.LaunchableID   // A (pod-wise) unique identifier for this launchable, used to distinguish it from other launchables in the pod
 	Version          string                // A version identifier
 	ServiceId        string                // A (host-wise) unique identifier for this launchable, used when creating runit services
 	RunAs            string                // The user to assume when launching the executable
@@ -46,7 +45,7 @@ type LaunchAdapter struct {
 	*Launchable
 }
 
-func (a LaunchAdapter) ID() manifest.LaunchableID {
+func (a LaunchAdapter) ID() launch.LaunchableID {
 	return a.Launchable.Id
 }
